Moves the setup_logger function to it's own file and move all the logger creation calls to the top of the modules instead on a per function/method basis.
