+++ CHANGELOG
@@ -302,3 +302,10 @@ v1.0.38
 v.1.0.39
   CHANGED: How endpoint are identified by "id" in config files. There was an
            issue with using "in" instead of "==" to mask the requested endpoints.
+
+v.1.0.40
+  CHANGED: Where the logger is generated by moving it to the top of the modules
+           instead of inside every function that needs it.
+  CHANGED: Moved the setup_logger and associated classes to their own 'logger.py'
+           file. This allowed to remove a bunch of logger creations buy just placing
+           the main one at the top.
