+++ dynafed_storagestats/base.py
@@ -9,6 +9,14 @@ import dynafed_storagestats.helpers
 import dynafed_storagestats.exceptions
 
 
+####################
+# Module Variables #
+####################
+
+# Creating logger
+_logger = logging.getLogger(__name__)
+
+
 ############
 # Classes ##
 ############
@@ -159,9 +167,6 @@ class StorageShare():
         Handles exceptions arising from missing/incorrect settings and logging.
 
         """
-        # Creating logger
-        _logger = logging.getLogger(__name__)
-
 
         _logger.info("[%s]Validating configured settings.", self.id)
 
@@ -266,9 +271,6 @@ class StorageShare():
         to deal with the different schema's to be checked.
 
         """
-        # Creating logger
-        _logger = logging.getLogger(__name__)
-
 
         _logger.debug(
             "[%s]Validating URN schema: %s",
