+++ dynafed_storagestats/version.py
@@ -1,3 +1,3 @@
 #!/usr/bin/env python3
 # -*- coding: UTF-8 -*-
-__version__ = "1.0.39"
+__version__ = "1.0.40"
